[timeline] Always use top frame for text location
Previously we were using the uiLocation's linkText, if existent, in
order to encode the location. Because of a backend bug
(https://crbug.com/1318635), we never were able to resolve the location
since at least ~April 2021. The code for using the linkText would,
however, also only work if we already had a uiLocation (which depends
on timing), and would be racy (if the bug hadn't been introduced).

This change makes the location text deterministic: use the top frame
as a link text. Note, that this is not a link, but just a text.

Bug: 1318635
Change-Id: I5078d18d92d8b60d8c8d4253efe07f7527bb3c19