Add a restricted override CSS variable to Button
We need to be able to change whether button has
right side border radius for some cases (i.e. measure performance button)

There are a few approaches to this:
* Having an override CSS variable for border-radius
* Having an explicit "Modifier" attribute for Button
* Having an override CSS variable like a boolean
that restricts the possible values for border-radius

-> First option opens up the component to be
more inconsistent in different places: i.e. someone
can set a border-radius with 8px which we don't
want for now

-> Second option bloats the API of the Button
and it can be confused with the `Variant`

-> Third option (this) can be best of the both
worlds:
* It restricts the usage of the override variable
to change border radius and still allows us to
have the presentation we want
* It doesn't bloat the API and doesn't make it possible
to confuse with Variant.

However, from the implementation perspective,
it also looks a bit hacky to use CSS variables this way
and it doesn't feel elegant.

Though it solves our issues well and answers the
questions we had. So, this is still a good alternative
from this regard.

Bug: 1306756
Change-Id: I702cf68374f6dab56e6fb0bde575268c326cc949