Deflake snippets e2e tests
The 'adds an expression to watches' was reenabled recently in a deflake attempt at https://crrev.com/c/3596912. However, the test remained flaky on the mac_relase bot. This is another attempt to deflake it. I ran 100 successful consecutive execs of the test with this CL's fix applied in the mac_release bot at https://logs.chromium.org/logs/devtools-frontend/buildbucket/cr-buildbucket/8815244297494067409/+/u/E2E_tests/stdout, so it is with greater confidence that it's deflaked this time

Bug: 1198160
Change-Id: Ib2d64de93a242cfa3caddae2b749ef50a35d2497