[WebAuthn] Add large blob toggle
Add a toggle that enables / disables large blob support for WebAuthn
virtual authenticators. This feature has been long supported by the
devtools protocol but we hadn't surfaced it on the front-end.

Large blobs are only available for resident-key enabled authenticators,
so we disable the checkbox if that option is not selected. Also, large
blobs require CTAP 2.1.

Fixed: 1321803
Change-Id: Ic1654276ba5a39ddd8943cd8689978117546e70e