[l10n] Fix deprecation issue messages
The CL https://crrev.com/c/3613872 applied some weird formatting to
the UIStrings object of DeprecationIssue.ts. This then caused the
strings to be no longer collected by our l10n infra (the CL removes
them from en-US.json).

This CL fixes that by turning off clang-format in that file and let
the PRESUBMIT script re-populate en-US.json with the UIStrings from
that file.

We have not yet imported these strings into the TC, so we were
lucky in catching this early.

Bug: None
Change-Id: I05b9faca446e19ff57ef3eccd04e919234797f8d