Add description for invalid attribution reporting header issue
The description here is vague, but because the issue includes both a reference to the original request and the invalid header value, it
should still be more useful to developers than the situation before this
CL, which was for invalid headers to be silently ignored.

We will add additional metadata, such as the name of the invalid header
and the reason why it was invalid, in future changes.

Screenshot of reported issue available in crbug.com/1302318#c4.

Bug: 1302318
Change-Id: Ie7f3b5e34b25ff56d832bf553ffcac674b5a57ba