Correct autocompletion for valid identifiers.
This regressed with https://crrev.com/c/3310606 where `SPAN_IDENT` was
introduced as a filter for properties to show in the autocompletion box
when using the dot MemberExpression syntax. The `SPAN_IDENT` regex looks
wrong and unnecessarily complicated.

This change uses the ID_Start and ID_Continue unicode property escapes
instead, so we get the right match here.

Fixed: 1324466
Bug: 1275553
Change-Id: I6f0d57771a7370ca761df8a62d488586992e9e3a