Add network request context menu entry for creating header override
This CL adds the ability to create header overrides from the list of
network requests in the network panel via the context menu.

If a local folder for storing local overrides has already been
configured beforehand, the context menu takes the user straight to the
'.headers' file where header overrides are to be specified. If
overrides have not been configured before, the user has to name a
local folder for storing overrides first.

Video: https://i.imgur.com/kLGlOtL.mp4

Bug: 1297533
Change-Id: I8b31fe41d922e1662da68320fcc188c596c0a058