Port General section of NetworkPanel's Headers View to custom component
Before adding an editor for response headers to the headers view, this
CL creates a custom component for the headers view. This is the first
of multiple CL's porting the existing headers view's functionality to
the new component.

The new component is only shown if the 'local overrides for response
headers' experiment is turned on.

Screenshot: https://i.imgur.com/qh4gL0k.png

Bug: 1297533
Change-Id: I2e49731546f4b4584626c61e886dcb1accafca8c