Adjust data grid visible elements based on header size in scroller
This fixes a regression due to
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3586197

The above change modified the data grid DOM so that the header row is a
child of the scrollable container. This meant that ViewportDataGrid's
update method which depends on the scrollable container's visible
viewport size needed to be adjusted to not count the header's height as
part of this visible area. To do this, we subtract the header height
from the scroll container's clientHeight to get the actual visible
height that can be used for data rows.

This works for most instances of ViewportDataGrid in the codebase, but
NetworkDataGrid sets a custom scroll container element using
setScrollContainer. This scroll container happens to already have the
correct final height; That is, its height is equal to the height
available for showing data rows. So when the header height is
subtracted from this, the ViewportDataGrid renders fewer rows than it
should, leaving a gap that the user can see.

To fix, I've updated the headerHeight() method to return zero when a
custom scroll container is being used. This assumes that since the
caller is using their own scroll container, then they are keeping its
height in sync with the visible data area themselves. So there is no
need for the update() method to perform any adjustment.

Bug: 1319308
Change-Id: I7184a6e00c3555a6ed07c1a6e68b997ffd776044