Jump to the to the assigned slot of an element
DISABLE_THIRD_PARTY_CHECK=need updated cdp for slots


It is now possible to jump to the assigned slot of an element.
Elements that have a slot have a button next to element
that reveals the assigned slot of this element.
Previously, it was possible to reveal which nodes are assigned
to a specific slot, but not the other way around.

Bug: none
Change-Id: I0913441c86ed5c2f89dbb8cac274f771896c7e73