[A11y] Fix DOMBreakpoint sidebar not announcing the check state
Screen Reader is not announcing the checkbox state in DOMbp sidebar. This is because the aria label is set at the checkbox element, however, the focusable element is its parent element. Now move the aria-label to it's parent so when the bp item get focused, the checkbox state will be announced.

Bug:1327440
Change-Id: I94fd9d3e0b93377f684fea6b5d384c944e29dc3d