[e2e] Report source files added to the navigator tree eagerly
Source files are reported to e2e tests through an event that's fired
when nodes in the navigator tree are expanded. This makes testing
unnecessarily hard if source files aren't in the same "directory",
because nodes won't autoexpand in that case. This CL reports files when
they're added to the tree instead.

Bug: 1328729
Change-Id: Ica41bcafb9e61e9bc73618364e6cb519a0ad57c1