Remove 1 usage of Protocol.Emulation.UserAgentMetadata
Protocol.Emulation.UserAgentMetadata has a superset of the properties we
need here, and eliminating usage of Protocol types is favorable because
CDP types are more likely to change than a web platform API like
WorkerNavigator.

WorkerNavigator.userAgentData is standardized but realtively new, so we
can make typescript happy by extending the Navigator interface to include
it.

Bug: none
Change-Id: I48b760f2329dd95ca6a7926e613449ec753fbd34