Expose web font `src` from protocol
This is useful for perf panels where we want to try to link from a node
for a metric (e.g. LCP) and highlight that a font was part of the cause
(e.g. slow network request = delayed LCP).

The protocol gives us the `src`, which will be a URL, but we were not
exposing it in the `CSSFontFace` class, until this CL.

Bug: 1326481
Change-Id: Ibde63ecd4df3eff91095ca48293c34b416015d5b