Await automapping and related status updates on instrumentation breaks
Breakpoints that are set on file system UISourceCodes will be moved
to their corresponding UISourceCode in the network project.

When paused on an instrumentation break we need to make sure that
all mappings (from file system UISourceCode -> network UISourceCode)
have already happened, before continuing and restoring breakpoints.
With this CL, we now actively wait for any binding updates, before
continuing with the restoration.

Bug: 1280621
Change-Id: I4d847e26bc27fb24c2b42c88cc347c447d964083