[sources] Use embedderName in ResourceMapping.
This changes the ResourceMapping to handle <script>s by their
embedderName (which is the URL under which they are known to
Blink) rather than by sourceURL, which can be affected by the
`//# sourceURL=...` annotations in the code.

This way the user can now properly set breakpoints in HTML
<script>s independent of whether there's a `//# sourceURL`
annotation or not. If such an annotation is present, the
breakpoint will automatically be set in the appropriate
standalone script (which will be automatically revealed).

Fixed: chromium:1329113
Change-Id: Ic1c098f7481188a642bc3fa597a1496efe9eff9e