Deflake and reenable styles pane e2e tests
It seems using keyboard presses to navigate the elements tree can
sometimes flake if the keyboard press happens before the tree is
ready. This change instead selects the element using text selectors
and clicks the element.
This change also changes some asserts to instead wait for the expected
styles to appear.

Bug: 1112692, 1318314, 1297458
Change-Id: I0709665252108cd457c17d1cc377f2fea0416e50