Account for content type when doing script deduplication
Content type includes whether the script is from source maps,
so this is sufficient to prevent a script from source maps from
conflicting with a deployed script.

ID for deduplication is now distinct from url for history tracking,
as we may want the former to change when updating the content but
not the latter.

Bug: 1335076
Change-Id: I72d916045b40adb999dccbe87f1c5a69feb1ee98