Top Layer elements representation element
This CL adds an ElementsTreeElements that contains links to elements in the top layer to a body tag. Developers can now use the elements panel to tell which elements are in the top layer and navigate between the top layer representation element and the top layer elements.

DISABLE_THIRD_PARTY_CHECK='need both'

Bug: None
Change-Id: Iddfa500a9c5ac321b9182a03d30b9fd2afecc9c3