[Lint] fix component naming bug
I had an issue where I'd copied/pasted a component definition but
forgotten to change the litTagName property. I then realised that this
rule wasn't detecting the case where we have duplicate `litTagName`
items and should be, so we now check if we've seen a tag before and
error if so. This ensures no duplicate tag names can exist in any single
file.

Bug: none
Change-Id: I2fe5bdfe3bee6a2d2f52e90f9b80e5c82737f416