Fix for devtools-frontend full checkout paths on Windows
Currently calculating the path to chromium's root directory returns an empty string on Windows because the path used to compare against the executable path is not normalized.

```
PATH_TO_EXECUTED_FILE.indexOf('src/third_party/devtools-frontend'); -> -1 on Windows
```

The fix is to use the normalizedPath variable to return the correct substring.

Change-Id: Ia12232f3746466f3257975d652342ac8b9ed2578