[debugger] Optimize building reverse source maps
This CL builds reverse source mappings eagerly after parsing the
source maps.  This removes the expensive lazy filtering of mappings
for the cost of efficient eager generation of the reverse mappings. On
a large example (VSCode with minification), this CL adds extra 150ms
to the 850ms worth of source map mappings parsing, but it reduces the
cost of agressive reverse map building from 1+ minute to ~5s.

If this is too much of a cost, we can only filter the reverse mappings
eagerly, but sort lazily - that would bring down the eager overhead
from 150ms to 50ms.

This is in preparation for smarter source mapping (see the linked
bug), which relies on the reverse maps more heavily.

Bug: chromium:1335338
Change-Id: Id45e8bdfd345cdb618886b8ab41fdbbc773a4a36