[debugger] Resolve scope view source map identifiers with punctuation
Currently, we only resolve exact identifier tokens with source maps.
However, some minifiers (esbuild, in particular) merge identifiers
with subsequent comma, paren or semicolon. In this CL, we try to
detect if the merging happened and ignore the punctuation when
resolving identifiers.

With this patch, the scope view shows deobfuscated identifiers
on vscode.dev page.

Bug: chromium:1335338
Change-Id: If9adba0c39db4fa3c6214c37a8cb7f9c245ad81d