Fix source map name resolution for constructors with a super call
The devtools parser (Acorn) refuses to parse super calls in
non-constructors.  This causes deobfuscation of variables fail in
constructors with super calls because we wrap scopes in ordinary
'function's. This patch wraps the scope in a class constructor
instead; this should work for both ordinary functions/methods and for
constructors.

Bug: chromium:1335825
Change-Id: Ie1ea1ed498e7c87f08060c1553cc612d55e7536a