[BreakpointManager] Refactor breakpoint setting code
This CL refactors the code surrounding updating the debugger.
Apart from factoring some code out for readability, the main
changes are the following:

1. We have one single function that schedules the update:

Breakpoint.updateModel() ->
ModelBreakpoint.scheduleUpdateInDebugger()`

scheduleUpdateInDebugger now has a return value which
is used to determine whether we should run it another time
since there are pending updates, or if an error occurred.

If an error occurred, it's the Breakpoint itself that
decides to remove itself and _NOT_ the ModelBreakpoint
anymore.

2. We use a mutex to guard the scheduleUpdateInDebugger

Previously we guarded that code using an updatePromise.
For readability, I have changed it to be guarded by a
promise. This has a few consequences: 1. we have
one additional await before we actually schedule the update,
which changes timing, 2. Awaiting the promise of the
update schedule will only guarantee awaiting changes
made at the time that we schedule it. Previously we
had a `hasPendingUpdate` which would be updated, and
it would apply all changes until no pending changes
were left. This should however, as far as I can see,
not be a problem.

Fixed: 1334930
Change-Id: I88588b4e85d710ce2f2554fa05423338b52c46f7