[liveedit] Remove obsolete stack trace result parameters
The CL https://crrev.com/c/3691850 deprecated result parameters of
Debugger#setScriptSource and changed the CDP method to always leave
them out.

This means we no longer have to handle them on the frontend side.
Moreover, as live edit can no longer change functions that are
on the stack (for now), we don't have to explicitly "step in"
to trigger frame restarting. The new version will handle that
automatically for us in the backend.

R=kimanh@chromium.org

Bug: 1334484
Change-Id: I7104a77ea36cdb7eadcbbe022104a78d9eae07f0