[debugging] Source map name resolution with id+equals tokens
If there is no name section in source maps, we infer names by
translating source positions of name tokens and inferring
the source name from the mapped token. Some minifiers (esbuild)
merge identifier tokens with subsequent punctuation token.
This CL introduces handling of an identifier token merged with
an assignment.

We also fix handling of nested scopes in the resolveScopeInObject
tests (the scope chain needs to be reversed).

Bug: chromium:1335338
Change-Id: I8069c7d5290370897a7db1447b0f1e1824d72bd4