Redesign the reveal element button and add a link icon to the top-layer adorner
The button that reveals elements, ex. top layer element and slot, has previously not looked like a button which could be misleading. An adorner with a link icon has been used to make it look more button-like

Also added a link icon to the top-layer element adorner. Top-layer element adorner is responsible for jumping to another element when it is clicked and for consistency we add 'link' icon to every adorner that links to another element.

Screenshot of both changes: https://imgur.com/gallery/XdylqKF

Bug: 1313690, 1066772
Change-Id: Ia3930c4342dfbbdea3cb5fbd1a33b5d5476dc481