[e2e] Deflake Console -> Issues workflow e2e test
The flaky behaviour occurred because of a race condition in the test
environment between the dispatch of the topFrameNavigation event and
the caching of issues. When the former happened after the latter, issues
are cleared out and no issues are shown in the Issues panel by the time
the test is run.

To work around this, a waitForFunction is implemented where the
navigation to the resource page is retried no issues are found right
before the test is executed.

I've run 500 successful executions locally with this fix in place.

Fixed: 1322534
Change-Id: I485c64dd63cc8fbcaeefafc989c437fc44401550