Reveal array size in the memory viewer
This CL implements basic array highlighting inside of the
LinearMemoryViewer when debugging C/C++ using the C/C++
extension. This CL introduces highlighting for a subset
of ValueNode variables represented as
StaticallyTypedValueNodes.

The current expected functionality includes:
- Highlighting a selected array in the viewer.
- Changing highlight to a newly inspected array.
- Resetting highlight when debugger pauses.

This CL also adds unittests aimed at testing the
re-higlighting when a new variable is inspected and
resetting when debugger is paused.

Demo: https://imgur.com/a/E8WqyLc

Bug: 1336568
Change-Id: I2a569a5971b23a74c630870d2d581c1b97ee024f