[debugger] Skip popover calculation for inactive editors
Without this change, popover calculation is activated for all tabs
because all the popover helpers (one for each tab) listen to mouse
events of the Sources panel.

This CL attaches the popover listeners to the appropriate editor, so
the mouse events are only processed by the popover helper of the
active editor. This helps with some hanging observed on large pages
(when multiple large sources are opened in the debugger).

Bug: chromium:1340133
Change-Id: I2efc4e94d375111e8cfb65d85c5ae7342d3102f5