[debugger] Do not hide popover too eagerly
If DevTools is showing a popover and the user hovers to another popover
request too quickly (before the old popover "times out"), then the new
popover might be cancelled by the timeout. To the user, it will look
like the popover does not work.

This patch avoids cancelling new popovers request on timeout.

Bug: chromium:1340120
Change-Id: I14dc502840aeb2a5653f3e1f9e815e4ab7d9345b