Show href='' rather than href=unknown in elements panel.
If the developer wrote code such as `<foo href="">`, then elements would
show this as:

```
<foo href(unknown)>
```

Whereas for `<a href="">` the elements panel correctly shows:

```
<a href>
```

The issue here was that for any `href` attribute, we would try to
linkify it, either via `<x-link>` (if the element is an anchor) or if
not via the Linkifier. But if the value is empty, there is no link to
linkify! So let's detect that case and simply render the attribute as
is, with no attempt to link it.

As a drive-by, I first thought I'd have to update `linkifyURL`, but I
didn't, but I wrote a few unit tests to document the current behaviour
of it.

Fixed: 1253949
Change-Id: I95c96288141c04bf41fe14e70116de4e7669dbd7