[Lint] fix pathing rules for inject_checkbox_styles
I'd made a mistake in the pathing logic for resolving imports; I needed
the _directory_ of the TS file being linted, not the filename itself.

I then also had to update all the tests as they contained incorrect
filepaths which caused me to miss the bug in the first place.

Bug: 1316297
Change-Id: Ie8303291b23d078a9ea1937d28ab3bef73deb258