[ESLint] lint front_end/third_party
In crrev.com/c/3748106 we noticed that a CL was able to land that
imported a file without the `.js` extension.

This is because `.eslintignore` listed `front_end/third_party`, so that
folder wasn't linted at all. This is an issue because whilst most of the
code in that directory is not authored by us, the entry points usually
are, and we want to apply our lint rules there.

This CL removes the blanket ignore, instead preferring to ignore
`third_party/X/package/`, and a few other special cases, ensuring that
devtools-frontend authored files in `third_party/` do get linted.

The CL updates the `es_modules_import` rule to not apply the import
checks to files in `third_party` (as they cannot apply the same
conventions we use in DevTools), but it will still check for file
extensions on imports.

The remaining changes are small stylistic changes to align to our ESLint
rules.

Fixed: 1342530
Change-Id: I4ce438499b8c9c77c3fe661576a58bad59d9b061