Add highlighting for FormattedValueNodes
Previously, highlighting was only supported for
`StaticallyTypedValueNode`s. This CL extends the support to
`FormattedValueNode`s. Now, all inspectable types should
support highlighting.

The highlighting for pointers and pointer-to-pointer
follows the behavior of the C/C++ DWARF Extension. That is,
inspecting a pointer highlights the pointed to value.
In the case of pointer-to-pointer, the highlighted value
is a pointer. To inspect a pointer-to-pointer and highlight
the underlying value directly, it would be necessary to
modify the Extension.

For testing, https://emscripten-dbg-stories.netlify.app/
now includes more debugging examples:
- Opening linear memory inspector example (DWARF)
- Opening linear memory inspector example with primitives
(DWARF)
- Opening linear memory inspector example with structs
(DWARF)

These are used by:
- clicking the desired link
- opening DevTools
- opening sources tab
- opening the `.cc` file
- setting a breakpoint on the line of interest
- reloading the webpage (either the reload button or CTRL+r)
- clicking the memory icon

Now you should see the memory inspector with a
highlighted byte range. Clicking on a different variable's
icon should change the highlight to that variable.

Bug: 1336568
Change-Id: Iab40b1e9586840b562f51aecd04867ce569e1eb4