Add highlighting for FormattedValueNodes
Previously, highlighting was only supported for
`StaticallyTypedValueNode`s. This CL extends the support to
`FormattedValueNode`s. Now, all inspectable types should
support highlighting.

The highlighting for pointers and pointer-to-pointer
follows the behavior of the C/C++ DWARF Extension. That is,
inspecting a pointer highlights the pointed to value.
In the case of pointer-to-pointer, the highlighted value
is a pointer. To inspect a pointer-to-pointer and highlight
the underlying value directly, it would be necessary to
modify the Extension.

The instructions below illustrate the changes in this
CL. The example focuses on a reference to the double type.

These are used by:
- head to https://emscripten-dbg-stories.netlify.app/
- open "Opening linear memory inspector with primitives"
- open DevTools
- open sources tab
- open `lmi-primitives.cc` in the left menu
- set a breakpoint on line 4
- reload the webpage (either the reload button or CTRL+r)
- click the memory icon next to `d: double &`

Now, you should see the memory inspector with a
highlighted byte range. This highlighted range should be
8 bytes, which is the size of a double variable.
Before this CL, only the first byte would be highlighted.

Before/After screenshot: https://imgur.com/a/lfQPBAD

Bug: 1336568
Change-Id: Iab40b1e9586840b562f51aecd04867ce569e1eb4